Add multiple file use checks; fix sed pattern

+++ unity-git.sh
@@ -25,20 +25,25 @@ fi
 
 # Mac Only!
 LFS_INSTALL="$(brew list git-lfs | grep bin/git-lfs )"
-echo $LFS_INSTALL
 if [[ $LFS_INSTALL = "Error*" || ! -f $LFS_INSTALL ]]; then
     echo "No git lfs, installing via brew..."
     brew install git-lfs
 fi
 
-curl -o .gitignore https://raw.githubusercontent.com/github/gitignore/master/Unity.gitignore
+if [ ! -f ".gitignore" ]; then
+    curl -o .gitignore https://raw.githubusercontent.com/github/gitignore/master/Unity.gitignore
+fi
 
 echo $'*.psd\n*.png\n*.mp3\n*.wav\n*.prefab\n*.fbx\n*.xcf\n*.blend\n*.obj\n*.3ds\n*.dae' > .gitattributes
-sed -e 's/\(\*\.\w\+\)/\1 filter=lfs diff=lfs merge=lfs -text\n/' .gitattributes > a.tmp && mv a.tmp .gitattributes
+sed -e 's/\(\*\.[a-zA-Z0-9]*\)/\1 filter=lfs diff=lfs merge=lfs -text/' .gitattributes > a.tmp && mv a.tmp .gitattributes
 
-printf $'[merge]\ntool = unityyamlmerge\n\n[mergetool "unityyamlmerge"]\ntrustExitCode = false\ncmd = \'' > .gitconfig && printf "$TOOL_PATH" >> .gitconfig && printf $'\' "$BASE" "$REMOTE" "$LOCAL" "$MERGED"\n' >> .gitconfig
+if [ ! -f ".gitconfig" ]; then
+    printf $'[merge]\ntool = unityyamlmerge\n\n[mergetool "unityyamlmerge"]\ntrustExitCode = false\ncmd = \'' > .gitconfig && printf "$TOOL_PATH" >> .gitconfig && printf $'\' "$BASE" "$REMOTE" "$LOCAL" "$MERGED"\n' >> .gitconfig
+fi
 
-git init
+if [ ! -d ".git" ]; then
+    git init
+fi
 git lfs install
 
 echo "~~Finished git setup for Unity~~~"
